sprint 1

+++ goss-testing/scripts/python/iuf_cleanup.py
@@ -24,41 +24,50 @@
 #
 
 """
-This script executes logs deletion for an iuf activity.
+This script executes deletion of logs , workflows and directories created for an iuf activity.
 """
 
 import subprocess
 import sys
+from pathlib import Path
 
-ACTIVITY_NAME = "test-activity"
 LOG_DIR= "/etc/cray/upgrade/csm/iuf"
 MEDIA_DIR= "/etc/cray/upgrade/csm/automation-tests"
 
-def get_workflows():
-    command = f"iuf -a {ACTIVITY_NAME} workflow"
+def get_workflows(activity_name):
+    command = f"iuf -a {activity_name} workflow"
     workflows=[]
     try:
         result = subprocess.run(command, shell=True, check=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE, universal_newlines=True)
-        print("Command output:", result.stdout)
-        workflows=result.stdout.splitlines()
-        
+        if not result.returncode:
+            workflows=result.stdout.splitlines()
     except subprocess.CalledProcessError as e:
-        print(f"Error: {e.stderr.strip()}")
         return e.returncode
-    
     return workflows
 
-def main():
-    command_delete_logs = f"rm -r {LOG_DIR}/{ACTIVITY_NAME}"
+def main(activity_name = "test-activity"):
+    command_delete_logs = f"rm -r {LOG_DIR}/{activity_name}"
     command_delete_media_dir = f"rm -r {MEDIA_DIR}"
     try:
-        workflows = get_workflows()
-        for workflow in workflows:
-            command_delete_workflow = f"kubectl delete workflow {workflow} -n argo"
-            result = subprocess.run(command_delete_workflow, shell=True, check=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE, universal_newlines=True)
-            print("Command output:", result.stdout)
-        result = subprocess.run(command_delete_logs, shell=True, check=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE, universal_newlines=True)
-        result = subprocess.run(command_delete_media_dir, shell=True, check=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE, universal_newlines=True)
+        workflows = get_workflows(activity_name)
+        if isinstance(workflows, list) and all(isinstance(wf, str) for wf in workflows) :
+            print(f"Workflows found for {activity_name} :{workflows}")
+            for workflow in workflows:
+                command_delete_workflow = f"kubectl delete workflow {workflow} -n argo"
+                result = subprocess.run(command_delete_workflow, shell=True, check=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE, universal_newlines=True)
+                print("Command output:", result.stdout)
+        logs_path = Path(f"{LOG_DIR}/{activity_name}")
+        if logs_path.is_dir():
+            print(f"{logs_path} exists.")
+            result = subprocess.run(command_delete_logs, shell=True, check=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE, universal_newlines=True)
+        else:
+            print(f"{logs_path} does not exist.")
+        media_path = Path(MEDIA_DIR)         
+        if media_path.is_dir():
+            print(f"{media_path} exists.")
+            result = subprocess.run(command_delete_media_dir, shell=True, check=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE, universal_newlines=True)
+        else:
+            print(f"{media_path} does not exist." )   
     except subprocess.CalledProcessError as e:
         print(f"Error: {e}")
         return e.returncode
@@ -66,5 +75,13 @@ def main():
     return 0
 
 if __name__ == "__main__":
-    exit_code = main()
-    sys.exit(exit_code)
+    if len(sys.argv) > 2:
+        print("Usage: script.py <ACTIVITY_NAME>")
+        sys.exit(1)
+    elif len(sys.argv) == 2:
+        activity_name = sys.argv[1]
+        exit_code = main(activity_name)
+        sys.exit(exit_code)
+    else:
+        exit_code = main()
+        sys.exit(exit_code)
